Tidy up initialization code, for readability
